# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Lealta 2.0 API
  description: |
    Sistema de fidelización multi-tenant con funcionalidades de IA y gestión avanzada de clientes.
    
    ## Características principales:
    - Sistema multi-tenant con aislamiento de datos
    - Autenticación robusta con control de roles
    - Sistema de tarjetas de lealtad automatizado
    - Integración con IA para procesamiento de tickets
    - APIs RESTful con validación completa
    
    ## Autenticación:
    La API utiliza autenticación basada en sesiones con cookies httpOnly.
    Roles disponibles: SUPERADMIN, ADMIN, STAFF
    
  version: 1.0.0
  contact:
    name: Lealta Development Team
    email: support@lealta.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api
    description: Desarrollo local
  - url: https://api.lealta.com
    description: Producción
paths:
...
  /staff/consumo:
    post:
      summary: Procesar ticket con IA
      description: Procesa un ticket usando OCR y Google Gemini AI
      tags:
        - Consumos
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
                - cedula
                - locationId
              properties:
                image:
                  type: string
                  format: binary
                  description: Imagen del ticket
                cedula:
                  type: string
                  description: Cédula del cliente
                locationId:
                  type: string
                  description: ID de la ubicación
                empleadoId:
                  type: string
                  description: ID del empleado (opcional)
      responses:
        '200':
          description: Ticket procesado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  requiresConfirmation:
                    type: boolean
                  empleadoDetectado:
                    type: string
                  total:
                    type: number
                  puntos:
                    type: number
                  productos:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProductoDetectado'
                  ocrText:
                    type: string
                  ticketImageUrl:
                    type: string
                  cliente:
                    type: object
                    properties:
                      nombre:
                        type: string
                      puntosActuales:
                        type: number
                      puntosNuevos:
                        type: number
  /staff/consumo/manual:
    post:
      summary: Registrar consumo manual
      description: Registra un consumo manualmente confirmado
      tags:
        - Consumos
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - cedula
                - empleadoVenta
                - productos
                - totalManual
              properties:
                cedula:
                  type: string
                empleadoVenta:
                  type: string
                productos:
                  type: array
                  items:
                    type: object
                    properties:
                      nombre:
                        type: string
                      cantidad:
                        type: number
                      precio:
                        type: number
                totalManual:
                  type: number
                ocrText:
                  type: string
                ticketImageUrl:
                  type: string
      responses:
        '200':
          description: Consumo registrado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  consumo:
                    $ref: '#/components/schemas/Consumo'
                  cliente:
                    $ref: '#/components/schemas/Cliente'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Autenticación basada en sesiones con cookies httpOnly
  schemas:
...
    Cliente:
      type: object
      properties:
        id:
          type: string
        cedula:
          type: string
        nombre:
          type: string
        telefono:
          type: string
        correo:
          type: string
          format: email
        puntos:
          type: number
        totalVisitas:
          type: number
        totalGastado:
          type: number
        fechaRegistro:
          type: string
          format: date-time
        tarjetaLealtad:
          $ref: '#/components/schemas/TarjetaLealtad'
...
    TarjetaLealtad:
      type: object
      properties:
        id:
          type: string
        nivel:
          type: string
          enum: [BRONCE, PLATA, ORO, PLATINO, DIAMANTE]
        activa:
          type: boolean
        fechaAsignacion:
          type: string
          format: date-time
        beneficios:
          type: array
          items:
            type: string
    Consumo:
      type: object
      properties:
        id:
          type: string
        clienteId:
          type: string
        monto:
          type: number
        puntos:
          type: number
        productos:
          type: array
          items:
            $ref: '#/components/schemas/ProductoDetectado'
        empleadoVenta:
          type: string
        createdAt:
          type: string
          format: date-time
    ProductoDetectado:
      type: object
      properties:
        nombre:
          type: string
        cantidad:
          type: number
        precio:
          type: number
...